Move test in test_atmPSF into test_psf with other psf tests
